pyproject.toml
[project]
name = "penport"
version = "0.1.0"
description = "Bridge handwritten journals to digital diaries via a local LLM pipeline"
requires-python = ">=3.11"
dependencies = [
    "wxPython>=4.2",
    "platformdirs>=4.0",
    "google-genai>=1.0",
    "openai>=1.0",
    "anthropic>=0.30",
    "requests>=2.31",
    "Pillow>=10.0",
]

[project.scripts]
penport = "main:main"

[dependency-groups]
dev = [
    "pyinstaller>=6.0",
    "black>=24.0",
    "ruff>=0.4",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
